refactor(ui): extract AppShell from Main.kt
Main.kt now is the pure 30-line process entry: uiModule definition
+ LauncherBootstrap.preBoot + PuppetServer start + application{}
calling AppShell(boot).

AppShell.kt holds the Compose-side root: AppState + Screen sealed
classes (UI navigation state), the window + tray + raise-tick
machinery, the migration/AppRoot branch, and AppRoot itself with
the auto-login LaunchedEffect (now delegating to
AutoLoginCoordinator) and the background + AppLayout render.

AppShell is an extension on ApplicationScope so exitApplication()
in the window close handler resolves without re-wiring the
receiver chain.

Main.kt: 822 -> 41 LOC. AppShell.kt: 568 LOC. The 213-LOC delta
went to launcher.bootstrap across the prior three commits in
this branch (X11/Display + LauncherBootstrap + AutoLoginCoordinator).
